Fix 3 mutation-suite hang bugs and add BenchmarkDotNet benchmark project
Phase 6.1: seeds 26346/27682/27724 hung for minutes in ChdFile.Open on corrupted
V5 headers. Root cause: totalbytes corruptions produced totalblocks up to ~3.5e9,
making compressed_v5_map allocate a multi-GB MapEntry[] and decode for minutes.
Bound the V5 map against its Huffman-RLE encoding capacity (max 91 entries/bit)
and int.MaxValue raw-map size before allocation; bound V1-V4 Totalblocks by the
inline map size that must physically fit in the file; reject cyclic SELF chains
in LinkSelfBlocks (previously recursive in ReadBlock). All 3501 mutation tests
now pass on net8/9/10 (~70s each).

Phase 7.4: add CHDSharpBench running BenchmarkDotNet 0.15.8 (offline cache) on
the repo corpus: per-codec decode throughput, Read byte-throughput, CheckFile
deep-verify cold/warm/with-parent, and per-codec encode MB/s (1 vs 8 workers,
MemoryDiagnoser). Add a --chdman comparer harness that runs chdman.exe verify/
createhd side-by-side with the library and prints a median wall-clock table.
Also includes working-tree style modernizations (enum PascalCase, switch expr,
await using, analyzer cleanups).

+++ CHDSharpLib/ChdHashing.cs
@@ -1,7 +1,6 @@
 using System.Diagnostics;
 using System.Security.Cryptography;
 using CHDSharp.Models;
-using CHDSharp.Utils;
 using System.IO.Hashing;
 
 namespace CHDSharp;
@@ -91,7 +90,7 @@ public static partial class Chd
                 foreach (var track in chd.Tracks)
                 {
                     var start = track.StartFrame * unitBytes;
-                    var length = (long)(track.Frames + track.ExtraFrames) * unitBytes;
+                    var length = (track.Frames + track.ExtraFrames) * unitBytes;
                     regions.Add((track.TrackNumber, start, length));
                 }
             }
